Update screaminterface.lisp
Some changes in the non deterministic listener to make it work with the modal dialog. 
For now it's working with score objects and with screamer functions, but it's only prints the correct result with command+click evaluation (otherwise always returns nil).
Some changes should be made in the size of the score objects to make it fits into the listener size.